% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_cleaning.R
\name{clean_team_abbr}
\alias{clean_team_abbr}
\title{Standardise NFL team abbreviations}
\usage{
clean_team_abbr(abbr)
}
\arguments{
\item{abbr}{a character vector of abbreviations}
}
\value{

}
\description{
This function standardises NFL team abbreviations to nflverse defaults.
This helps for joins and plotting, especially with the new nflplotR package!
}
\examples{
{
x <- c("LVR","PIE","LAC","PIT")
clean_team_abbr(x)
}
}
