% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_cleaning.R
\name{clean_team_abbrs}
\alias{clean_team_abbrs}
\title{Standardise NFL team abbreviations}
\usage{
clean_team_abbrs(abbr, current_location = TRUE)
}
\arguments{
\item{abbr}{a character vector of abbreviations}
}
\value{

}
\description{
This function standardises NFL team abbreviations to nflverse defaults.
This helps for joins and plotting, especially with the new nflplotR package!
}
\examples{
{
x <- c("PIE","LAR","PIT","CRD", "OAK", "CLV")
clean_team_abbr(x)
}
}
