% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cache_clear.R
\name{clear_cache}
\alias{clear_cache}
\alias{.clear_cache}
\title{Clear function cache}
\usage{
clear_cache()

.clear_cache()
}
\value{
A success message after clearing the cache.
}
\description{
This function clears the memoised cache of all functions memoised by \code{nflreadr}.
}
\examples{
clear_cache()
}
