% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_combine}
\alias{dictionary_combine}
\title{Data Dictionary: ESPN QBR}
\format{
An object of class \code{data.frame} with 18 rows and 3 columns.
}
\usage{
dictionary_combine
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_combine]{load_combine()}}
}
\seealso{
\code{vignette("Data Dictionary - Combine")}

\url{https://nflreadr.nflverse.com/articles/dictionary_combine.html}
}
\keyword{datasets}
