% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_contracts}
\alias{dictionary_contracts}
\title{Data Dictionary: Contracts}
\format{
An object of class \code{data.frame} with 50 rows and 3 columns.
}
\usage{
dictionary_contracts
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_contracts]{load_contracts()}}
}
\seealso{
\code{vignette("Data Dictionary - Contracts")}

\url{https://nflreadr.nflverse.com/articles/dictionary_contracts.html}
}
\keyword{datasets}
