% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_depth_charts}
\alias{dictionary_depth_charts}
\title{Data Dictionary: Depth Charts}
\format{
An object of class \code{data.frame} with 13 rows and 3 columns.
}
\usage{
dictionary_depth_charts
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_depth_charts]{load_depth_charts()}}
}
\seealso{
\code{vignette("Data Dictionary - Depth Charts")}

\url{https://nflreadr.nflverse.com/articles/dictionary_depth_charts.html}
}
\keyword{datasets}
