% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_draft_picks}
\alias{dictionary_draft_picks}
\title{Data Dictionary: Draft Picks}
\format{
An object of class \code{data.frame} with 10 rows and 3 columns.
}
\usage{
dictionary_draft_picks
}
\description{
A dataframe containing the data dictionary for \code{load_draft_picks()}
}
\seealso{
\code{vignette("Data Dictionary - Draft Picks")}
}
\keyword{datasets}
