% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_espn_qbr}
\alias{dictionary_espn_qbr}
\title{Data Dictionary: ESPN QBR}
\format{
An object of class \code{data.frame} with 23 rows and 3 columns.
}
\usage{
dictionary_espn_qbr
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_espn_qbr]{load_espn_qbr()}}
}
\seealso{
\code{vignette("Data Dictionary - ESPN QBR")}

\url{https://nflreadr.nflverse.com/articles/dictionary_espn_qbr.html}
}
\keyword{datasets}
