% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_ff_opportunity}
\alias{dictionary_ff_opportunity}
\title{Data Dictionary: Expected Fantasy Points}
\format{
An object of class \code{data.frame} with 219 rows and 4 columns.
}
\usage{
dictionary_ff_opportunity
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_ff_opportunity]{load_ff_opportunity()}}
}
\seealso{
\code{vignette("Data Dictionary - Expected Fantasy Points")}

\url{https://nflreadr.nflverse.com/articles/dictionary_ff_opportunity.html}
}
\keyword{datasets}
