% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_ff_playerids}
\alias{dictionary_ff_playerids}
\title{Data Dictionary: Fantasy Player IDs}
\format{
An object of class \code{data.frame} with 34 rows and 3 columns.
}
\usage{
dictionary_ff_playerids
}
\description{
A dataframe containing the data dictionary for \code{load_ff_playerids()}
}
\seealso{
\code{vignette("Data Dictionary - FF Player IDs")}
}
\keyword{datasets}
