% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_ff_rankings}
\alias{dictionary_ff_rankings}
\title{Data Dictionary: Fantasy Football Rankings}
\format{
An object of class \code{data.frame} with 25 rows and 3 columns.
}
\usage{
dictionary_ff_rankings
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_ff_rankings]{load_ff_rankings()}}
}
\seealso{
\code{\link{vignette("Data Dictionary - FF Rankings")}}
}
\keyword{datasets}
