% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_injuries}
\alias{dictionary_injuries}
\title{Data Dictionary: Injuries}
\format{
An object of class \code{data.frame} with 16 rows and 3 columns.
}
\usage{
dictionary_injuries
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_injuries]{load_injuries()}}
}
\seealso{
\code{vignette("Data Dictionary - Injuries")}

\url{https://nflreadr.nflverse.com/articles/dictionary_injuries.html}
}
\keyword{datasets}
