% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_nextgen_stats}
\alias{dictionary_nextgen_stats}
\title{Data Dictionary: Next Gen Stats}
\format{
An object of class \code{data.frame} with 61 rows and 3 columns.
}
\usage{
dictionary_nextgen_stats
}
\description{
A dataframe containing the data dictionary for \code{load_nextgen_stats()}
}
\seealso{
\code{vignette("Data Dictionary - Next Gen Stats")}
}
\keyword{datasets}
