% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_participation}
\alias{dictionary_participation}
\title{Data Dictionary: Participation}
\format{
An object of class \code{data.frame} with 14 rows and 3 columns.
}
\usage{
dictionary_participation
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_participation]{load_participation()}}
}
\seealso{
\code{vignette("Data Dictionary - Participation")}

\url{https://nflreadr.nflverse.com/articles/dictionary_participation.html}
}
\keyword{datasets}
