% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_pbp}
\alias{dictionary_pbp}
\title{Data Dictionary: Play by Play}
\format{
An object of class \code{data.frame} with 372 rows and 2 columns.
}
\usage{
dictionary_pbp
}
\description{
A dataframe containing the data dictionary for \code{load_pbp()}
}
\seealso{
\code{vignette("dictionary_pbp")}
}
\keyword{datasets}
