% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_pfr_passing}
\alias{dictionary_pfr_passing}
\title{Data Dictionary: PFR Passing}
\format{
An object of class \code{data.frame} with 28 rows and 3 columns.
}
\usage{
dictionary_pfr_passing
}
\description{
A dataframe containing the data dictionary for \code{load_pfr_passing()}
}
\seealso{
\code{vignette("Data Dictionary - PFR Passing")}
}
\keyword{datasets}
