% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_player_stats}
\alias{dictionary_player_stats}
\title{Data Dictionary: Player Stats}
\format{
An object of class \code{data.frame} with 41 rows and 2 columns.
}
\usage{
dictionary_player_stats
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_player_stats]{load_player_stats()}}
}
\seealso{
\code{vignette("Data Dictionary - Player Stats")}

\url{https://nflreadr.nflverse.com/articles/dictionary_player_stats.html}
}
\keyword{datasets}
