% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_player_stats_def}
\alias{dictionary_player_stats_def}
\title{Data Dictionary: Player Stats Defense}
\format{
An object of class \code{data.frame} with 22 rows and 3 columns.
}
\usage{
dictionary_player_stats_def
}
\description{
A dataframe containing the data dictionary for \code{\link{load_player_stats(stat_type = "defense")}}
}
\seealso{
\code{vignette("Data Dictionary - Player Stats Defense")}

\url{https://nflreadr.nflverse.com/articles/dictionary_player_stats_def.html}
}
\keyword{datasets}
