% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_players}
\alias{dictionary_players}
\title{Data Dictionary: Players}
\format{
An object of class \code{data.frame} with 39 rows and 3 columns.
}
\usage{
dictionary_players
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_players]{load_players()}}
}
\seealso{
\code{vignette("Data Dictionary - Players")}

\url{https://nflreadr.nflverse.com/articles/dictionary_players.html}
}
\keyword{datasets}
