% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_roster_status}
\alias{dictionary_roster_status}
\title{dictionary_roster_status}
\format{
An object of class \code{data.frame} with 19 rows and 2 columns.
}
\usage{
dictionary_roster_status
}
\description{
A dictionary translating the shorthand roster status to more verbose explanations of what each status indicates.
}
\keyword{datasets}
