% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_rosters}
\alias{dictionary_rosters}
\title{Data Dictionary: Rosters}
\format{
An object of class \code{data.frame} with 24 rows and 3 columns.
}
\usage{
dictionary_rosters
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_rosters]{load_rosters()}}
}
\seealso{
\code{vignette("Data Dictionary - Rosters")}

\url{https://nflreadr.nflverse.com/articles/dictionary_rosters.html}
}
\keyword{datasets}
