% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_schedules}
\alias{dictionary_schedules}
\title{Data Dictionary: Schedules}
\format{
An object of class \code{data.frame} with 29 rows and 2 columns.
}
\usage{
dictionary_schedules
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_schedules]{load_schedules()}}
}
\seealso{
\code{vignette("Data Dictionary - Schedules")}

\url{https://nflreadr.nflverse.com/articles/dictionary_schedules.html}
}
\keyword{datasets}
