% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_snap_counts}
\alias{dictionary_snap_counts}
\title{Data Dictionary: Snap Counts}
\format{
An object of class \code{data.frame} with 12 rows and 3 columns.
}
\usage{
dictionary_snap_counts
}
\description{
A dataframe containing the data dictionary for \code{load_snap_counts()}
}
\seealso{
\code{vignette("Data Dictionary - Snap Counts")}
}
\keyword{datasets}
