% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_team_stats}
\alias{dictionary_team_stats}
\title{Data Dictionary: Team Stats}
\format{
An object of class \code{data.frame} with 102 rows and 3 columns.
}
\usage{
dictionary_team_stats
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_team_stats]{load_team_stats()}}
}
\seealso{
\code{vignette("Data Dictionary - Team Stats")}

\url{https://nflreadr.nflverse.com/articles/dictionary_team_stats.html}
}
\keyword{datasets}
