% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_trades}
\alias{dictionary_trades}
\title{Data Dictionary: Trades}
\format{
An object of class \code{data.frame} with 11 rows and 3 columns.
}
\usage{
dictionary_trades
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_trades]{load_trades()}}
}
\seealso{
\code{vignette("Data Dictionary - Trades")}

\url{https://nflreadr.nflverse.com/articles/dictionary_trades.html}
}
\keyword{datasets}
