% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cran.R
\name{.for_cran}
\alias{.for_cran}
\title{CRAN setup code}
\usage{
.for_cran()
}
\description{
Some code to comply with CRAN policies, e.g.
limiting data.table CPU cores for examples/tests
}
\keyword{internal}
