% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sitrep.R
\name{.sitrep_pkg_deps}
\alias{.sitrep_pkg_deps}
\title{Show dependency versions of installed packages}
\usage{
.sitrep_pkg_deps(
  packages,
  inst_pkgs = utils::installed.packages(),
  recursive = TRUE
)
}
\description{
Show dependency versions of installed packages
}
\keyword{internal}
