% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sitrep.R
\name{.sitrep_pkg_opts}
\alias{.sitrep_pkg_opts}
\title{Show options for installed packages}
\usage{
.sitrep_pkg_opts(packages, redact_path = TRUE)
}
\description{
Show options for installed packages
}
\keyword{internal}
