% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-piggyback.R
\name{download_nflverse}
\alias{download_nflverse}
\title{Bulk download utilities via piggyback}
\usage{
download_nflverse(
  ...,
  folder_path = getOption("nflreadr.download_path", default = "."),
  file_type = getOption("nflreadr.prefer", default = "rds")
)
}
\arguments{
\item{...}{releases to download, provided in either unquoted or character format
(i.e. pbp or "pbp" are both fine)}

\item{folder_path}{a folder in which subfolders will be created for each release -
defaults to path specified in \verb{options("nflreadr.download_path)} or "." (the current working directory)}

\item{file_type}{one of \code{c("rds","parquet", "csv", "qs")} -
defaults to file type specified in \code{options(nflreadr.prefer)} or "rds"}
}
\description{
This function downloads or updates data from the nflverse-data repository releases,
creating subfolders that match the release structure.
}
\examples{
{
\donttest{
try({
  temp_dir <- tempdir()

  download_nflverse(combine, contracts, folder_path = temp_dir, file_type = "parquet")

  unlink(temp_dir) # cleanup
})
}
}
}
