% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-arrow.R
\name{download_parquet}
\alias{download_parquet}
\title{Load Arrow Dataset}
\usage{
download_parquet(dir, data_type, seasons = TRUE, overwrite = TRUE)
}
\arguments{
\item{dir}{directory to load data to}

\item{data_type}{requested data frame, corresponds to location of thing.}

\item{seasons}{seasons for data frame, defaults to TRUE and loads all data.}

\item{overwrite}{overwrite existing files/directories? default = TRUE}
}
\value{
a success message and the supplied path where the data folder was created (invisibly)
}
\description{
This function creates subfolders within the specified directory and downloads parquet files such that you can work with them via \code{arrow::open_dataset()}.
}
\examples{
\donttest {
  try({
    download_parquet(
      dir = tempdir(),
      data_type = "pbp",
      seasons = TRUE,
      overwrite = TRUE
    )
  })
}
}
