% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_date.R
\name{most_recent_season}
\alias{most_recent_season}
\alias{get_latest_season}
\alias{get_current_season}
\title{Get Latest Season}
\usage{
most_recent_season(roster = FALSE)

get_latest_season(roster = FALSE)

get_current_season(roster = FALSE)
}
\arguments{
\item{roster}{Either \code{TRUE} or \code{FALSE}.
If \code{TRUE}, will return current year after March 15th, otherwise previous year.
If \code{FALSE}, will return current year on or after Thursday following Labor Day,
i.e. Thursday after the first Monday in September. Otherwise previous year.}
}
\value{
most recent season (a four digit numeric)
}
\description{
A helper function to choose the most recent season available for a given dataset
}
\seealso{
Other Date utils: 
\code{\link{get_current_week}()}
}
\concept{Date utils}
