% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nflverse.R
\name{load_depth_charts}
\alias{load_depth_charts}
\title{Load Weekly Depth Charts}
\usage{
load_depth_charts(seasons = most_recent_season())
}
\arguments{
\item{seasons}{a numeric vector specifying what seasons to return, if \code{TRUE} returns all available data}
}
\value{
A tibble of week-level depth charts for each team.
}
\description{
Loads depth charts for each NFL team for each week back to 2001.
}
\examples{
\donttest{
  load_depth_charts(2020)
}

}
\seealso{
\url{https://github.com/nflverse/nflfastR-roster}
}
