% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nflverse.R
\name{load_draft_picks}
\alias{load_draft_picks}
\title{Load Draft Picks from PFR}
\usage{
load_draft_picks()
}
\value{
A tibble of NFL draft picks provided by Pro Football Reference.
}
\description{
Loads every draft pick since 1980 courtesy of PFR.
}
\examples{
\donttest{
  load_draft_picks()
}

}
