% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nflverse.R
\name{load_espn_qbr}
\alias{load_espn_qbr}
\title{Load ESPN's QBR}
\usage{
load_espn_qbr(
  league = c("nfl", "college"),
  seasons = most_recent_season(),
  summary_type = c("season", "weekly"),
  file_type = "csv"
)
}
\arguments{
\item{league}{One of "nfl" or "college"}

\item{seasons}{a numeric vector of seasons to return, data available since 2006. Defaults to latest season available. TRUE will select all seasons.}

\item{summary_type}{One of "season" or "weekly", defaults to season}

\item{file_type}{currently csv, soon rds/qs options will be available!}
}
\value{
a tibble of season-level injury report data.
}
\description{
Load ESPN's QBR
}
\examples{
\donttest{
    load_espn_qbr("nfl",2020)
}

}
\seealso{
\url{https://github.com/nflverse/nflfastR-roster}
}
