% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ffverse.R
\name{load_ff_playerids}
\alias{load_ff_playerids}
\title{Load Fantasy Player IDs}
\usage{
load_ff_playerids()
}
\value{
a dataframe of player IDs
}
\description{
Accesses DynastyProcess.com's database of fantasy football player IDs, which help connect nflverse to various other platforms and IDs.
}
\examples{
\donttest{
load_ff_playerids()
}

}
\seealso{
\url{https://github.com/dynastyprocess/data}

\code{\link[ffscrapr:dp_playerids]{ffscrapr::dp_playerids()}}
}
