% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nflverse.R
\name{load_pfr_passing}
\alias{load_pfr_passing}
\title{Load Advanced Passing Stats from PFR}
\usage{
load_pfr_passing(seasons = TRUE)
}
\arguments{
\item{seasons}{a numeric vector specifying what seasons to return, if \code{TRUE} returns all available data}
}
\value{
A tibble of season-level player statistics provided by Pro Football Reference.
}
\description{
Loads player level season stats provided by Pro Football Reference
starting with the 2019 season.
}
\examples{
\donttest{
  load_pfr_passing()
}

}
\seealso{
\url{https://www.pro-football-reference.com/years/2020/passing_advanced.htm}
}
