% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_players.R
\name{load_players}
\alias{load_players}
\title{Load Players}
\usage{
load_players()
}
\value{
A tibble with one row per player.
}
\description{
Load a dataframe of player-level information, including IDs and other mostly-immutable data (birthdates, college, draft position etc.)
}
\examples{
\donttest{
try({ # prevents cran errors
  load_players()
})
}

}
\seealso{
Issues with this data should be filed here: \url{https://github.com/nflverse/nflreadr} and it will be triaged appropriately.
}
