% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nflverse.R
\name{load_team_graphics}
\alias{load_team_graphics}
\title{Load Team Graphics}
\usage{
load_team_graphics()
}
\value{
A tibble of team-level image URLs and hex colour codes.
}
\description{
Loads team graphics, colours, and logos - useful for plots!
}
\examples{
\donttest{
  load_team_graphics()
}

}
\seealso{
\url{https://github.com/nflverse/nflfastr-data}
}
