% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_teams.R
\name{load_teams}
\alias{load_teams}
\title{Load NFL Team Graphics, Colors, and Logos}
\usage{
load_teams(
  current = TRUE,
  file_type = getOption("nflreadr.prefer", default = "rds")
)
}
\arguments{
\item{current}{If \code{TRUE} (the default), returns a standardized list of current teams only,
with abbreviations as per \link{team_abbr_mapping}.}

\item{file_type}{One of \code{c("rds", "qs", "csv", "parquet")}. Can also be set globally with
\code{options(nflreadr.prefer)}}
}
\value{
A tibble of team-level image URLs and hex color codes.
}
\description{
Loads team graphics, colors, and logos - useful for plots!
}
\examples{
\dontshow{.for_cran()}
\donttest{
try({ # prevents cran errors
  load_teams()
})
}

}
\seealso{
Issues with this data should be filed here: \url{https://github.com/nflverse/nflverse-pbp}
}
