% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_trades.R
\name{load_trades}
\alias{load_trades}
\title{Load Trades}
\usage{
load_trades(seasons = TRUE)
}
\arguments{
\item{seasons}{a numeric vector of seasons to return, default \code{TRUE} returns all available data.}
}
\value{
A tibble of game information for past and/or future games.
}
\description{
This returns a table of historical trades as maintained by Lee Sharpe.
}
\examples{
\donttest{
 load_trades(2020)
}

}
\seealso{
\url{https://nflreadr.nflverse.com/articles/dictionary_trades.html} for a web version of the dictionary

\code{\link{dictionary_trades}} for the data dictionary as bundled within the package

Issues with this data should be filed here: \url{https://github.com/nflverse/nfldata}
}
