% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nflverse_data_timezone}
\alias{nflverse_data_timezone}
\title{nflverse Timezone}
\format{
An object of class \code{character} of length 1.
}
\usage{
nflverse_data_timezone
}
\description{
A character string defining the default timezone for data across the nflverse
}
\keyword{datasets}
