% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflverse_game_id.R
\name{nflverse_game_id}
\alias{nflverse_game_id}
\title{Compute nflverse Game Identifiers}
\usage{
nflverse_game_id(season, week, away, home)
}
\arguments{
\item{season}{4 digit season between 1999 and the output of \code{\link[=most_recent_season]{most_recent_season()}}}

\item{week}{Numeric or character giving the week, bewtween 1 and 22.}

\item{home, away}{Valid NFL team abbreviation as it can be found in \link{team_abbr_mapping}}
}
\value{
A character vector
}
\description{
Compute nflverse Game Identifiers
}
\examples{
nflverse_game_id(2022, 2, "LAC", "KC")
}
