% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download_nflverse.R
\name{nflverse_releases}
\alias{nflverse_releases}
\title{List all available nflverse releases}
\usage{
nflverse_releases()
}
\value{
A dataframe containing release names, release descriptions, and
other relevant release information.
}
\description{
This functions lists all nflverse data releases that are available
in the nflverse-data repo. Release names can be used for downloads in
\code{\link[=nflverse_download]{nflverse_download()}}.
}
\examples{
\donttest{
# Change option for better output
old <- options(piggyback.verbose = FALSE)

try( # avoids cran failures, can skip in normal usage
nflverse_releases()
)

# Restore old options
options(old)
}
}
