% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sitrep.R
\name{nflverse_sitrep-class}
\alias{nflverse_sitrep-class}
\title{nflverse_sitrep data class}
\description{
\code{nflverse_sitrep()} returns an S3 class that was created to allow for custom
printing. It will otherwise dispatch to the \code{list} class.
}
\details{
Re-exporting with \code{methods::setOldClass()} allows these dataframes to be used
in S4 generics like those from DBI and jsonlite.
}
\keyword{internal}
