% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_url.R
\name{parquet_from_url}
\alias{parquet_from_url}
\title{Load .parquet file from a remote connection}
\usage{
parquet_from_url(url)
}
\arguments{
\item{url}{a character url}
}
\value{
a dataframe as parsed by \code{\link[arrow:read_parquet]{arrow::read_parquet()}}
}
\description{
Load .parquet file from a remote connection
}
\examples{
\donttest{
parquet_from_url(
"https://github.com/nflverse/nflverse-data/releases/download/player_stats/player_stats.parquet"
)
}
}
