% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{player_name_mapping}
\alias{player_name_mapping}
\title{Alternate player name mappings}
\format{
A named character vector
\describe{
\item{name attribute}{The "alternate" name.}
\item{value attribute}{The "correct" name.}
}
}
\usage{
player_name_mapping
}
\description{
A named character vector mapping common alternate names, re-exported from \code{ffscrapr}.
}
\details{
You can suggest additions to this table by \href{https://github.com/ffverse/ffscrapr/issues/new/choose}{opening an issue in ffscrapr}.
}
\examples{
\donttest{
player_name_mapping[c("Chatarius Atwell", "Robert Kelley")]
}

}
\keyword{datasets}
