% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_url.R
\name{progressively}
\alias{progressively}
\title{Progressively}
\usage{
progressively(f, p = NULL)
}
\arguments{
\item{f}{a function}

\item{p}{a progressor}
}
\value{
the same function as \code{f} but it calls \code{p()} on exit, which is suitable for progressr calls
}
\description{
Progressively
}
\examples{
{progressively(rds_from_url)}

}
