% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_url.R
\name{qs_from_url}
\alias{qs_from_url}
\title{Load .qs file from a remote connection}
\usage{
qs_from_url(url)
}
\arguments{
\item{url}{a character url}
}
\value{
a dataframe
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

The underlying package qs has been removed from CRAN on 2026-01-17 so
nflreadr stopped the qs support
}
\examples{
\dontshow{.for_cran()}
\donttest{
try({
  qs_from_url(
    "https://github.com/nflverse/nflverse-data/releases/download/player_stats/player_stats.qs"
  )
})
}
}
\keyword{internal}
