% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_url.R
\name{raw_from_url}
\alias{raw_from_url}
\title{Load raw filedata from a remote connection}
\usage{
raw_from_url(url, ...)
}
\arguments{
\item{url}{a character url}

\item{...}{for internal usage only}
}
\description{
This function allows you to retrieve data from a URL into raw format, which
can then be passed into the appropriate file-reading function, such as \code{arrow::read_parquet()}
}
\examples{
\donttest{
raw_from_url(
"https://github.com/nflverse/nflfastR-data/raw/master/data/play_by_play_2020.parquet"
)
}
}
