% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbindlist_with_attrs}
\alias{rbindlist_with_attrs}
\title{rbindlist but maintain attributes of last file}
\usage{
rbindlist_with_attrs(dflist)
}
\description{
rbindlist but maintain attributes of last file
}
\keyword{internal}
