% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{team_abbr_mapping}
\alias{team_abbr_mapping}
\title{Alternate team abbreviation mappings}
\format{
A named character vector
\describe{
\item{name attribute}{The "alternate" name.}
\item{value attribute}{The "correct" name.}
}
}
\usage{
team_abbr_mapping
}
\description{
A named character vector mapping common alternate team abbreviations.
}
\details{
You can suggest additions to this table by \href{https://github.com/nflverse/nflreadr/issues/new/choose}{opening an issue in nflreadr}.
}
\examples{
\donttest{
team_abbr_mapping[c("STL", "OAK","CRD","BLT", "CLV")]
}

}
\seealso{
\code{team_abbr_mapping_norelocate} for the same thing but relocations stay in their original cities.
}
\keyword{datasets}
