% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{team_abbr_mapping_norelocate}
\alias{team_abbr_mapping_norelocate}
\title{Alternate team abbreviation mappings, no relocation}
\format{
A named character vector
\describe{
\item{name attribute}{The "alternate" name.}
\item{value attribute}{The "correct" name.}
}
}
\usage{
team_abbr_mapping_norelocate
}
\description{
A named character vector mapping common alternate team abbreviations, but does not follow relocations to their current city.
}
\details{
You can suggest additions to this table by \href{https://github.com/nflverse/nflreadr/issues/new/choose}{opening an issue in nflreadr}.
}
\examples{
\donttest{
team_abbr_mapping_norelocate[c("STL", "OAK","CRD","BLT", "CLV")]
}

}
\keyword{datasets}
