% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teams_colors_logos}
\alias{teams_colors_logos}
\title{NFL Team names, colors and logo urls.}
\format{
A data frame with 36 rows and 14 variables containing NFL team level
information, including franchises in multiple cities.
The colors are taken from Lee Sharpe's teamcolors.csv who has taken them from the
\code{teamcolors} package created by Ben Baumer and Gregory Matthews.
The Wikipeadia logo urls are taken from Lee Sharpe's logos.csv
Team wordmarks from nfl.com
}
\usage{
teams_colors_logos
}
\description{
NFL Team names, colors and logo urls.
}
\examples{
\donttest{
teams_colors_logos
}
}
\seealso{
\code{vignette("Data Dictionary - Teams Colors Logos")}
}
\keyword{datasets}
