% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_draft_picks}
\alias{dictionary_draft_picks}
\title{Data Dictionary: Draft Picks}
\format{
An object of class \code{data.frame} with 36 rows and 3 columns.
}
\usage{
dictionary_draft_picks
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_draft_picks]{load_draft_picks()}}
}
\seealso{
\code{vignette("Data Dictionary - Draft Picks")}

\url{https://nflreadr.nflverse.com/articles/dictionary_draft_picks.html}
}
\keyword{datasets}
