% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_ff_playerids}
\alias{dictionary_ff_playerids}
\title{Data Dictionary: Fantasy Player IDs}
\format{
An object of class \code{data.frame} with 35 rows and 3 columns.
}
\usage{
dictionary_ff_playerids
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_ff_playerids]{load_ff_playerids()}}
}
\seealso{
\code{vignette("Data Dictionary - FF Player IDs")}

\url{https://nflreadr.nflverse.com/articles/dictionary_ff_playerids.html}
}
\keyword{datasets}
