% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ftn_charting.R
\docType{data}
\name{dictionary_ftn_charting}
\alias{dictionary_ftn_charting}
\title{Data Dictionary: FTN Charting Data}
\format{
An object of class \code{data.frame} with 28 rows and 5 columns.
}
\usage{
dictionary_ftn_charting
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_ftn_charting]{load_ftn_charting()}}
}
\seealso{
\code{vignette("Data Dictionary - FTN Charting")}

\url{https://nflreadr.nflverse.com/articles/dictionary_ftn_charting.html}

Other ftn_charting: 
\code{\link{load_ftn_charting}()}
}
\concept{ftn_charting}
\keyword{datasets}
