% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nflverse.R
\name{load_combine}
\alias{load_combine}
\title{Load Combine Data from PFR}
\usage{
load_combine(seasons = TRUE)
}
\arguments{
\item{seasons}{a numeric vector of seasons to return, default \code{TRUE} returns all available data}
}
\value{
A tibble of NFL combine data provided by Pro Football Reference.
}
\description{
Loads combine data since 2000 courtesy of PFR.
}
\examples{
\donttest{
  load_combine()
}

}
\seealso{
Issues with this data should be filed here: \url{https://github.com/nflverse/nflfastR-data}

\url{https://nflreadr.nflverse.com/articles/dictionary_combine.html} for a web version of the dictionary

\code{\link{dictionary_combine}} for the data dictionary as bundled within the package

Issues with this data should be filed here: \url{https://github.com/nflverse/nfldata}
}
