% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ffverse.R
\name{load_ff_playerids}
\alias{load_ff_playerids}
\title{Load Fantasy Player IDs}
\usage{
load_ff_playerids()
}
\value{
a dataframe of player IDs
}
\description{
Accesses DynastyProcess.com's database of fantasy football player IDs, which help connect nflverse to various other platforms and IDs.
}
\examples{
\donttest{
try({ # prevents cran errors
load_ff_playerids()
})
}

}
\seealso{
\url{https://nflreadr.nflverse.com/articles/dictionary_ff_playerids.html} for the web data dictionary

Issues with this data should be filed here: \url{https://github.com/dynastyprocess/data}
}
