% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ffverse.R
\name{load_ff_rankings}
\alias{load_ff_rankings}
\title{Load Latest FantasyPros Rankings}
\usage{
load_ff_rankings(type = c("draft", "week", "all"))
}
\arguments{
\item{type}{one of "draft" (preseason), "week" (this week, inseason), or "all" (full archive)}
}
\value{
a dataframe of expert consensus rankings
}
\description{
Accesses DynastyProcess.com's repository of the latest FP expert consensus rankings - updated on a weekly basis.
}
\examples{
\donttest{
try({ # prevents cran errors
load_ff_rankings()
})
}

}
\seealso{
\url{https://nflreadr.nflverse.com/articles/dictionary_ff_rankings.html} for the web data dictionary

\url{https://www.fantasypros.com} for the source of data

Issues with this data should be filed here: \url{https://github.com/dynastyprocess/data}
}
