% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ffverse.R
\name{load_ff_rankings}
\alias{load_ff_rankings}
\title{Load Latest FantasyPros Rankings}
\usage{
load_ff_rankings()
}
\value{
a dataframe of expert consensus rankings
}
\description{
Accesses DynastyProcess.com's repository of the latest FP expert consensus rankings - updated on a weekly basis.
}
\examples{
\donttest{
load_ff_rankings()
}

}
\seealso{
\url{https://github.com/dynastyprocess/data}

\url{https://www.fantasypros.com}
}
