% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nflverse.R
\name{load_injuries}
\alias{load_injuries}
\title{Load Injury Reports}
\usage{
load_injuries(seasons = most_recent_season())
}
\arguments{
\item{seasons}{a numeric vector of seasons to return, defaults to returning
this year's data if it is March or later. If set to \code{TRUE}, will return all available data.}
}
\value{
a tibble of season-level injury report data.
}
\description{
Load Injury Reports
}
\examples{
\donttest{
    load_injuries(2020)
}

}
\seealso{
\url{https://github.com/nflverse/nflfastR-roster}
}
