% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nflverse.R
\name{load_injuries}
\alias{load_injuries}
\title{Load Injury Reports}
\usage{
load_injuries(
  seasons = most_recent_season(),
  file_type = getOption("nflreadr.prefer", default = "qs")
)
}
\arguments{
\item{seasons}{a numeric vector of seasons to return, data available since 2009. Defaults to latest season available.}

\item{file_type}{One of \code{"rds"} or \code{"qs"}. Can also be set globally with options(nflreadr.prefer)}
}
\value{
a tibble of season-level injury report data.
}
\description{
Load Injury Reports
}
\examples{
\donttest{
    load_injuries(2020)
}

}
\seealso{
\url{https://github.com/nflverse/nflfastR-roster}
}
