% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_injuries.R
\name{load_injuries}
\alias{load_injuries}
\title{Load Injury Reports}
\usage{
load_injuries(seasons = most_recent_season(), file_type = NULL)
}
\arguments{
\item{seasons}{a numeric vector of seasons to return, data available since 2009. Defaults to latest season available.}

\item{file_type}{Deprecated: now uses rds by default.}
}
\value{
a tibble of season-level injury report data.
}
\description{
Data collected from an API for weekly injury report data.
}
\examples{
\donttest{
    load_injuries(2020)
}

}
\seealso{
\url{https://nflreadr.nflverse.com/articles/dictionary_injuries.html} for a web version of the dictionary

\code{\link{dictionary_injuries}} for the data dictionary as bundled within the package

Issues with this data should be filed here: \url{https://github.com/nflverse/nflverse-data}
}
