% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_pbp}
\alias{load_pbp}
\title{Load Play By Play}
\usage{
load_pbp(seasons, file_type = c("rds", "parquet", "qs"))
}
\arguments{
\item{seasons}{A vector of 4-digit years associated with given NFL seasons.}

\item{file_type}{One of \code{"rds"}, \code{"parquet"} or \code{"qs"}. The latter two require
the package \code{curl} as well as \code{arrow} and \code{qs} respectively.}
}
\value{
The complete nflfastR dataset as returned by \code{\link[nflfastR:build_nflfastR_pbp]{nflfastR::build_nflfastR_pbp()}} for
all given \code{seasons}
}
\description{
Loads multiple seasons from the nflfastR data repository
}
\examples{
\donttest{
  load_pbp(2019:2020)
}

}
\seealso{
\url{https://www.nflfastr.com/articles/field_descriptions.html} for the data dictionary
}
