% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nflverse.R
\name{load_pfr_passing}
\alias{load_pfr_passing}
\title{Load Advanced Passing Stats from PFR}
\usage{
load_pfr_passing(seasons = TRUE)
}
\arguments{
\item{seasons}{a numeric vector specifying what seasons to return, if \code{TRUE} returns all available data}
}
\value{
A tibble of season-level player statistics provided by Pro Football Reference.
}
\description{
Loads player level season stats provided by Pro Football Reference
starting with the 2019 season.
}
\examples{
\donttest{
  load_pfr_passing()
}

}
\seealso{
\url{https://nflreadr.nflverse.com/articles/dictionary_pfr_passing.html} for the web data dictionary

\code{\link{dictionary_pfr_passing}} for the data dictionary as bundled within the package

Issues with this data should be filed here: \url{https://github.com/nflverse/pfr_scrapR}
}
