% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_player_stats}
\alias{load_player_stats}
\title{Load Player Level Weekly Stats}
\usage{
load_player_stats(seasons, file_type = c("rds", "qs", "parquet"))
}
\arguments{
\item{seasons}{a numeric vector of seasons to return}

\item{file_type}{one of "rds", "qs" or "parquet"}
}
\description{
Load Player Level Weekly Stats
}
