% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_player_stats}
\alias{load_player_stats}
\title{Load Player Level Stats}
\usage{
load_player_stats(seasons, type, summarise_by)
}
\arguments{
\item{seasons}{a numeric vector of seasons to return}

\item{type}{one of \code{offense}, \code{defense}, or \code{special_teams}}

\item{summarise_by}{one of \code{week} or \code{season}}
}
\description{
Load Player Level Stats
}
