% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_schedules}
\alias{load_schedules}
\title{Load Game/Schedule Data}
\usage{
load_schedules(seasons = TRUE)
}
\arguments{
\item{seasons}{a numeric vector of seasons to return, default \code{TRUE} returns all available data.}
}
\value{
A tibble of game information for past and/or future games.
}
\description{
This returns game/schedule information as maintained by Lee Sharpe.
}
\examples{
\donttest{
 load_schedules(2020)
}

}
\seealso{
\url{https://github.com/nflverse/nfldata}
}
