% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nflverse.R
\name{load_snap_counts}
\alias{load_snap_counts}
\title{Load Snap Counts from PFR}
\usage{
load_snap_counts(seasons = most_recent_season())
}
\arguments{
\item{seasons}{a numeric vector specifying what seasons to return, if \code{TRUE} returns all available data}
}
\value{
A tibble of game-level snap counts provided by Pro Football Reference.
}
\description{
Loads game level snap counts stats provided by Pro Football Reference
starting with the 2013 season.
}
\examples{
\donttest{
  load_snap_counts()
}

}
\seealso{
\url{https://nflreadr.nflverse.com/articles/dictionary_snap_counts.html} for the web data dictionary

\code{\link{dictionary_snap_counts}} for the data dictionary as bundled within the package
}
