% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_team_stats}
\alias{load_team_stats}
\title{Load Team Level Stats}
\usage{
load_team_stats()
}
\arguments{
\item{seasons}{a numeric vector of seasons to return}

\item{summarise_by}{one of \code{week} or \code{season}}

\item{type}{one of \code{offense}, \code{defense}, or \code{special_teams}}
}
\description{
Load Team Level Stats
}
