% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nflverse.R
\name{load_teams}
\alias{load_teams}
\title{Load NFL Team Graphics, Colors, and Logos}
\usage{
load_teams()
}
\value{
A tibble of team-level image URLs and hex color codes.
}
\description{
Loads team graphics, colors, and logos - useful for plots!
}
\examples{
\donttest{
  load_teams()
}

}
\seealso{
\url{https://github.com/nflverse/nflfastr-data}
}
