% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nflverse.R
\name{load_trades}
\alias{load_trades}
\title{Load Trades}
\usage{
load_trades(seasons = TRUE)
}
\arguments{
\item{seasons}{a numeric vector of seasons to return, default \code{TRUE} returns all available data.}
}
\value{
A tibble of game information for past and/or future games.
}
\description{
This returns a table of historical trades as maintained by Lee Sharpe.
}
\examples{
\donttest{
 load_trades(2020)
}

}
\seealso{
\url{https://github.com/nflverse/nfldata/blob/master/DATASETS.md#trades}

Issues with this data should be filed here: \url{https://github.com/nflverse/nfldata}
}
