% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{nflreadr-package}
\alias{nflreadr}
\alias{nflreadr-package}
\title{nflreadr: Download 'nflverse' Data}
\description{
A low-level package for downloading data from 'GitHub' repositories of the 'nflverse' project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nflreadr.nflverse.com}
  \item \url{https://github.com/nflverse/nflreadr}
  \item Report bugs at \url{https://github.com/nflverse/nflreadr/issues}
}

}
\author{
\strong{Maintainer}: Tan Ho \email{tan@tanho.ca} (\href{https://orcid.org/0000-0001-8388-5155}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Sebastian Carl \email{mrcaseb@gmail.com}
}

Other contributors:
\itemize{
  \item John Edwards \email{edwards1860@gmail.com} [contributor]
  \item Ben Baldwin \email{bbaldwin206@gmail.com} [contributor]
  \item Thomas Mock [contributor]
  \item Lee Sharpe [contributor]
}

}
\keyword{internal}
