% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download_nflverse.R
\name{nflverse_releases}
\alias{nflverse_releases}
\title{List all Available nflverse Releases}
\usage{
nflverse_releases()
}
\value{
A data.table containing release names and release descriptions
}
\description{
This functions lists all nflverse data releases that are available
in the nflverse-data repo. Release names can be used for downloads in
\code{\link[=download_nflverse]{download_nflverse()}}.
}
\examples{
\donttest{
try(
nflverse_releases()
)
}
}
