% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_url.R
\name{parquet_from_url}
\alias{parquet_from_url}
\title{Load .parquet file from a remote connection}
\usage{
parquet_from_url(url, ...)
}
\arguments{
\item{url}{a character url}

\item{...}{for internal usage only}
}
\value{

}
\description{
Load .parquet file from a remote connection
}
\examples{
parquet_from_url("https://github.com/nflverse/nflfastR-data/blob/master/data/play_by_play_2020.parquet?raw=true")
}
