% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_url.R
\name{qs_from_url}
\alias{qs_from_url}
\title{Load .qs file from a remote connection}
\usage{
qs_from_url(url, ...)
}
\arguments{
\item{url}{a character url}

\item{...}{for internal usage only}
}
\description{
Load .qs file from a remote connection
}
\examples{
qs_from_url(
"https://github.com/nflverse/nflfastR-data/blob/master/data/play_by_play_2020.qs?raw=true"
)
}
