% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_url.R
\name{raw_from_url}
\alias{raw_from_url}
\title{Load raw filedata from a remote connection}
\usage{
raw_from_url(url)
}
\arguments{
\item{url}{a character url}
}
\value{
a raw vector
}
\description{
This function allows you to retrieve data from a URL into raw format, which
can then be passed into the appropriate file-reading function. Data is memoised/cached for 24 hours.
}
\examples{
\donttest{
try({ # prevents CRAN errors
head(raw_from_url(
  "https://github.com/nflverse/nflverse-data/releases/download/test/combines.rds"
  ),
50)
})
}
}
