% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_url.R
\name{rds_from_url}
\alias{rds_from_url}
\title{Load .rds file from a remote connection}
\usage{
rds_from_url(url)
}
\arguments{
\item{url}{a character url}
}
\value{
a dataframe as created by \code{\link[=readRDS]{readRDS()}}
}
\description{
Load .rds file from a remote connection
}
\examples{
\donttest{
try({ # prevents cran errors
  rds_from_url("https://github.com/nflverse/nflverse-data/releases/download/test/combines.rds")
})
}
}
